fix: update ft_strncmp function signature to use size_t for the length parameter

+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: jhoban <jhoban@student.42berlin.de>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/17 21:17:05 by jhoban            #+#    #+#             */
-/*   Updated: 2025/11/24 10:56:49 by jhoban           ###   ########.fr       */
+/*   Updated: 2025/11/25 15:42:41 by jhoban           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ int		ft_isalnum(int c);
 int		ft_isprint(int c);
 void	*ft_memset(void *str, int c, size_t n);
 size_t	ft_strlen(const char *str);
-int		ft_strncmp(const char *s1, const char *s2, unsigned int n);
+int		ft_strncmp(const char *s1, const char *s2, size_t n);
 char	*ft_strchr(const char *s, int c);
 char	*ft_strrchr(const char *s, int c);
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len);
